Rename handlers to api
