+++ pkg/server/.gitignore
@@ -6,11 +6,3 @@ test-dnote
 /dist
 /build
 server
-
-# Elastic Beanstalk Files
-/tmp
-application.zip
-test-api
-/dump
-api
-/build
