+++ pkg/server/api/testutils.go
@@ -16,7 +16,7 @@
  * along with Dnote.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package handlers
+package api
 
 import (
 	"net/http/httptest"
