+++ pkg/server/api/v3_books.go
@@ -16,7 +16,7 @@
  * along with Dnote.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package handlers
+package api
 
 import (
 	"encoding/json"
