+++ pkg/server/handlers/helpers.go
@@ -1,83 +1,50 @@
-/* Copyright (C) 2019, 2020 Monomax Software Pty Ltd
- *
- * This file is part of Dnote.
- *
- * Dnote is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Dnote is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with Dnote.  If not, see <https://www.gnu.org/licenses/>.
- */
-
 package handlers
 
 import (
 	"encoding/json"
 	"net/http"
 	"strings"
+	"time"
 
-	"github.com/dnote/dnote/pkg/server/database"
 	"github.com/dnote/dnote/pkg/server/log"
-	"github.com/jinzhu/gorm"
 	"github.com/pkg/errors"
 )
 
-func paginate(conn *gorm.DB, page int) *gorm.DB {
-	limit := 30
-
-	// Paginate
-	if page > 0 {
-		offset := limit * (page - 1)
-		conn = conn.Offset(offset)
-	}
-
-	conn = conn.Limit(limit)
-
-	return conn
+// RespondForbidden responds with forbidden
+func RespondForbidden(w http.ResponseWriter) {
+	http.Error(w, "forbidden", http.StatusForbidden)
 }
 
-func getBookIDs(books []database.Book) []int {
-	ret := []int{}
-
-	for _, book := range books {
-		ret = append(ret, book.ID)
-	}
-
-	return ret
+// RespondUnauthorized responds with unauthorized
+func RespondUnauthorized(w http.ResponseWriter) {
+	UnsetSessionCookie(w)
+	w.Header().Add("WWW-Authenticate", `Bearer realm="Dnote Pro", charset="UTF-8"`)
+	http.Error(w, "unauthorized", http.StatusUnauthorized)
 }
 
-func validatePassword(password string) error {
-	if len(password) < 8 {
-		return errors.New("Password should be longer than 8 characters")
-	}
-
-	return nil
+// RespondNotFound responds with not found
+func RespondNotFound(w http.ResponseWriter) {
+	http.Error(w, "not found", http.StatusNotFound)
 }
 
-func getClientType(r *http.Request) string {
-	origin := r.Header.Get("Origin")
+// RespondInvalidSMTPConfig responds with invalid SMTP config error
+func RespondInvalidSMTPConfig(w http.ResponseWriter) {
+	http.Error(w, "SMTP is not configured", http.StatusInternalServerError)
+}
 
-	if strings.HasPrefix(origin, "moz-extension://") {
-		return "firefox-extension"
+// UnsetSessionCookie unsets the session cookie
+func UnsetSessionCookie(w http.ResponseWriter) {
+	expire := time.Now().Add(time.Hour * -24 * 30)
+	cookie := http.Cookie{
+		Name:     "id",
+		Value:    "",
+		Expires:  expire,
+		Path:     "/",
+		HttpOnly: true,
 	}
 
-	if strings.HasPrefix(origin, "chrome-extension://") {
-		return "chrome-extension"
-	}
-
-	userAgent := r.Header.Get("User-Agent")
-	if strings.HasPrefix(userAgent, "Go-http-client") {
-		return "cli"
-	}
-
-	return "web"
+	w.Header().Set("Cache-Control", "no-cache")
+	http.SetCookie(w, &cookie)
 }
 
 // HandleError logs the error and responds with the given status code with a generic status text
@@ -97,8 +64,8 @@ func HandleError(w http.ResponseWriter, msg string, err error, statusCode int) {
 	http.Error(w, statusText, statusCode)
 }
 
-// respondJSON encodes the given payload into a JSON format and writes it to the given response writer
-func respondJSON(w http.ResponseWriter, statusCode int, payload interface{}) {
+// RespondJSON encodes the given payload into a JSON format and writes it to the given response writer
+func RespondJSON(w http.ResponseWriter, statusCode int, payload interface{}) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(statusCode)
 
@@ -107,27 +74,80 @@ func respondJSON(w http.ResponseWriter, statusCode int, payload interface{}) {
 	}
 }
 
-// notSupported is the handler for the route that is no longer supported
-func (a *API) notSupported(w http.ResponseWriter, r *http.Request) {
+// NotSupported is the handler for the route that is no longer supported
+func NotSupported(w http.ResponseWriter, r *http.Request) {
 	http.Error(w, "API version is not supported. Please upgrade your client.", http.StatusGone)
 	return
 }
 
-func respondForbidden(w http.ResponseWriter) {
-	http.Error(w, "forbidden", http.StatusForbidden)
+// getSessionKeyFromCookie reads and returns a session key from the cookie sent by the
+// request. If no session key is found, it returns an empty string
+func getSessionKeyFromCookie(r *http.Request) (string, error) {
+	c, err := r.Cookie("id")
+
+	if err == http.ErrNoCookie {
+		return "", nil
+	} else if err != nil {
+		return "", errors.Wrap(err, "reading cookie")
+	}
+
+	return c.Value, nil
 }
 
-func respondUnauthorized(w http.ResponseWriter) {
-	unsetSessionCookie(w)
-	w.Header().Add("WWW-Authenticate", `Bearer realm="Dnote Pro", charset="UTF-8"`)
-	http.Error(w, "unauthorized", http.StatusUnauthorized)
+type authHeader struct {
+	scheme     string
+	credential string
 }
 
-// RespondNotFound responds with not found
-func RespondNotFound(w http.ResponseWriter) {
-	http.Error(w, "not found", http.StatusNotFound)
+func parseAuthHeader(h string) (authHeader, error) {
+	parts := strings.Split(h, " ")
+
+	if len(parts) != 2 {
+		return authHeader{}, errors.New("Invalid authorization header")
+	}
+
+	parsed := authHeader{
+		scheme:     parts[0],
+		credential: parts[1],
+	}
+
+	return parsed, nil
 }
 
-func respondInvalidSMTPConfig(w http.ResponseWriter) {
-	http.Error(w, "SMTP is not configured", http.StatusInternalServerError)
+// getSessionKeyFromAuth reads and returns a session key from the Authorization header
+func getSessionKeyFromAuth(r *http.Request) (string, error) {
+	h := r.Header.Get("Authorization")
+	if h == "" {
+		return "", nil
+	}
+
+	payload, err := parseAuthHeader(h)
+	if err != nil {
+		return "", errors.Wrap(err, "parsing the authorization header")
+	}
+	if payload.scheme != "Bearer" {
+		return "", errors.New("unsupported scheme")
+	}
+
+	return payload.credential, nil
+}
+
+// GetCredential extracts a session key from the request from the request header. Concretely,
+// it first looks at the 'Cookie' and then the 'Authorization' header. If no credential is found,
+// it returns an empty string.
+func GetCredential(r *http.Request) (string, error) {
+	ret, err := getSessionKeyFromCookie(r)
+	if err != nil {
+		return "", errors.Wrap(err, "getting session key from cookie")
+	}
+	if ret != "" {
+		return ret, nil
+	}
+
+	ret, err = getSessionKeyFromAuth(r)
+	if err != nil {
+		return "", errors.Wrap(err, "getting session key from Authorization header")
+	}
+
+	return ret, nil
 }
